Seperate core/web test projects
