+++ src/ImageProcessor.Web.Plugins.PostProcessor/Properties/AssemblyInfo.cs
@@ -9,6 +9,7 @@
 // --------------------------------------------------------------------------------------------------------------------
 
 using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
@@ -42,3 +43,5 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 [assembly: AssemblyVersion("1.2.3.0")]
 [assembly: AssemblyFileVersion("1.2.3.0")]
+
+[assembly: InternalsVisibleTo("ImageProcessor.Web.UnitTests")]
