+++ tests/ImageProcessor.Web.UnitTests/ImageProcessor.Web.UnitTests.csproj
@@ -52,6 +52,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Caching\AsyncKeyLockTests.cs" />
+    <Compile Include="Caching\DiskCacheTests.cs" />
     <Compile Include="ExtendedColorTypeConverterTests.cs" />
     <Compile Include="Extensions\DirectoryInfoExtensionsUnitTests.cs" />
     <Compile Include="Extensions\StringExtensionsUnitTests.cs" />
